% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_store_creds.R
\name{sf_store_creds}
\alias{sf_store_creds}
\title{Install SalesForce credentials in Your \code{.Renviron} File for Repeated Use}
\usage{
sf_store_creds(
  username = NULL,
  password = NULL,
  token = NULL,
  overwrite = FALSE,
  install = FALSE
)
}
\arguments{
\item{username}{The email address you use to login to SalesForce formatted in quotes.}

\item{password}{The password you use to login to SalesForce formatted in quotes.}

\item{token}{The API token provided to you from SalesForce formatted in quotes. A key can be acquired as described here \url{https://help.salesforce.com/s/articleView?id=sf.user_security_token.htm&type=5}}

\item{overwrite}{If this is set to TRUE, it will overwrite the existing variables that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install your credentials in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\value{

}
\description{
This function will add your username, password, and SalesForce
API token to your \code{.Renviron} file so it can be called securely without
being stored in your code. After you have installed your variables, they
the variables can be called any time by typing \code{Sys.getenv("SF_USER")},
\code{Sys.getenv("SF_PASSWORD")}, or \code{Sys.getenv("SF_TOKEN")}. If you do
not have an \code{.Renviron} file, the function will create on for you.
If you already have an \code{.Renviron} file, the function will append the
variables to your existing file, while making a backup of your original file
for disaster recovery purposes.

You can update all of your variables at once, or update them one at a time.
}
