% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfForce-package.R
\docType{package}
\name{stfForce-package}
\alias{stfForce}
\alias{stfForce-package}
\title{stfForce: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
